packages/backend/requirements.txt
llama-index
llama-index-core
llama-index-embeddings-openai
pinecone-client
llama-index-vector-stores-pinecone
llama-index-graph-stores-neo4j
python-dotenv
flask
boto3
urllib3==1.26.7
requests
beautifulsoup4